feat(config): simplify ansible python package management
This commit streamlines the Ansible configuration by removing the
redundant `geerlingguy.pip` role and `role_pip_packages` variable.
Instead, all necessary Python packages are now installed directly
using the `package` module. This simplifies the playbook and
reduces the number of moving parts, making the configuration easier
to understand and maintain. Additionally the portainer stack file
is updated to use the correct jinja2 syntax, and terraform and
homepage are updated to support access tokens.